# api_exposing/openapi.yaml
openapi: 3.0.3
info:
  title: Api VeRecipes
  description: |-
    This API exposes data from Vegetarian and Vegan Recipes collected from several the websites Jamie Oliver, Veganuary and The Simple Veganist
    
    
  contact:
    email: your@email.here
  license:
    name: CC BY-NC 3.0
    url: https://creativecommons.org/licenses/by-nc/3.0/
  version: 1.0.0
paths:
  /websites/:
    get:
      summary: Get all websites IDs and names
      description: Return all websites IDs and names
      responses:
        '200':
          description: OK          
        '404':
          description: Website not found
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
      security:
        - api_auth: []

  /recipes:
    get:
      summary: Get a list of all recipes available from all the websites
      description: Return the list with recipes.
      parameters:
        - name: page
          in: query
          example: 0
          description: Index of the page (used for pagination)
          schema:
            type: integer
        - name: page_size
          in: query
          example: 100
          description: Number of recipes per page (used for pagination)
          schema:
            type: integer
        - name: website_id
          in: query
          example: 4
          description: The website ID (used to gather a list of recipes from a specific website)
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  movies:
                    type: array
                    items:
                      $ref: '#'
                  next_page:
                    type: integer
                    example: 0
                  last_page:
                    type: integer
                    example: 100
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
      security:
        - api_auth: []

  /recipes/{id}:
    get:
      summary: Get a recipe by its ID
      description: Return all the information about a recipe - ingredients list, steps to prepare it ect.
      parameters:
        - name: id
          in: path
          description: ID of the recipe
          example: VN10
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK       
        '404':
          description: recipe not found
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
      security:
        - api_auth: []